Code has been refactored so that the javascript and vue are different. README updated as search functionality is now complete.
